mcp-examples/requirements.txt
# Core
python-dotenv==1.2.1

# LangGraph and LangChain
langgraph==1.0.4
langchain-core==1.1.1
langchain-openai==1.1.0
langchain-mcp-adapters>=0.2.1
